vmm: Move the VM creation and startup helpers to the api module
They're API wrappers, not VMM ones.
